+++ .env.example
@@ -1,5 +1,15 @@
 # # Environment Variables
-# The following variables are used to configure the production environment.
+
+APP_ENV=development
+
+NEXT_PUBLIC_API_URL="http://localhost:8080/v1"
+NEXT_PUBLIC_SITE_TITLE="02Beta"
+NEXT_PUBLIC_SITE_DESCRIPTION="02Beta is a platform for creating and managing your business."
+NEXT_PUBLIC_PRIVACY_URL="https://02beta.com/privacy"
+NEXT_PUBLIC_TERMS_URL="https://02beta.com/terms"
+NEXT_PUBLIC_COOKIES_URL="https://02beta.com/cookies"
+NEXT_PUBLIC_IMPRINT_URL="https://02beta.com/imprint"
+
 # ------------------------------------------------------------
 # ## POSTGRES DATABASE
 # ------------------------------------------------------------
