+++ apps/api/routes/auth/router.py
@@ -124,7 +124,7 @@ async def login(
     except Exception as e:
         raise HTTPException(
             status_code=status.HTTP_401_UNAUTHORIZED,
-            detail=f"Invalid credentials: {e}",
+            detail=str(e),
         )
