+++ apps/web/app/(auth)/login/login-form.tsx
@@ -3,7 +3,7 @@
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import * as z from "zod";
-import { Button } from "../ui/button";
+import { Button } from "@workspace/ui/components/ui/button";
 import {
   Form,
   FormControl,
@@ -11,47 +11,46 @@ import {
   FormItem,
   FormLabel,
   FormMessage,
-} from "../ui/form";
-import { Input } from "../ui/input";
+} from "@workspace/ui/components/ui/form";
+import { Input } from "@workspace/ui/components/ui/input";
 import { toast } from "sonner";
 import { useState } from "react";
 import { Loader2 } from "lucide-react";
 
-const resetPasswordSchema = z
-  .object({
-    password: z.string().min(8, "Password must be at least 8 characters"),
-    confirmPassword: z.string(),
-  })
-  .refine((data) => data.password === data.confirmPassword, {
-    message: "Passwords don't match",
-    path: ["confirmPassword"],
-  });
+const loginSchema = z.object({
+  email: z.string().email("Invalid email address"),
+  password: z
+    .string()
+    .min(8, "Password must be at least 8 characters")
+    .regex(
+      /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z0-9]).{8,}$/,
+      "Password must contain at least one uppercase letter, one lowercase letter, one number, and one special character",
+    ),
+});
 
-type ResetPasswordFormValues = z.infer<typeof resetPasswordSchema>;
+type LoginFormValues = z.infer<typeof loginSchema>;
 
-interface ResetPasswordFormProps {
-  onSubmit: (values: ResetPasswordFormValues) => Promise<void>;
+interface LoginFormProps {
+  onSubmit: (values: LoginFormValues) => Promise<void>;
 }
 
-export function ResetPasswordForm({ onSubmit }: ResetPasswordFormProps) {
+export function LoginForm({ onSubmit }: LoginFormProps) {
   const [isLoading, setIsLoading] = useState(false);
 
-  const form = useForm<ResetPasswordFormValues>({
-    resolver: zodResolver(resetPasswordSchema),
+  const form = useForm<LoginFormValues>({
+    resolver: zodResolver(loginSchema),
     defaultValues: {
+      email: "",
       password: "",
-      confirmPassword: "",
     },
   });
 
-  async function handleSubmit(values: ResetPasswordFormValues) {
+  async function handleSubmit(values: LoginFormValues) {
     setIsLoading(true);
     try {
       await onSubmit(values);
     } catch (error) {
-      toast.error(
-        error instanceof Error ? error.message : "Failed to reset password",
-      );
+      toast.error(error instanceof Error ? error.message : "Login failed");
     } finally {
       setIsLoading(false);
     }
@@ -62,12 +61,12 @@ export function ResetPasswordForm({ onSubmit }: ResetPasswordFormProps) {
       <form onSubmit={form.handleSubmit(handleSubmit)} className="space-y-4">
         <FormField
           control={form.control}
-          name="password"
+          name="email"
           render={({ field }) => (
             <FormItem>
-              <FormLabel>New Password</FormLabel>
+              <FormLabel>Email</FormLabel>
               <FormControl>
-                <Input placeholder="••••••••" type="password" {...field} />
+                <Input placeholder="you@example.com" type="email" {...field} />
               </FormControl>
               <FormMessage />
             </FormItem>
@@ -75,10 +74,10 @@ export function ResetPasswordForm({ onSubmit }: ResetPasswordFormProps) {
         />
         <FormField
           control={form.control}
-          name="confirmPassword"
+          name="password"
           render={({ field }) => (
             <FormItem>
-              <FormLabel>Confirm Password</FormLabel>
+              <FormLabel>Password</FormLabel>
               <FormControl>
                 <Input placeholder="••••••••" type="password" {...field} />
               </FormControl>
@@ -88,7 +87,7 @@ export function ResetPasswordForm({ onSubmit }: ResetPasswordFormProps) {
         />
         <Button type="submit" className="w-full" disabled={isLoading}>
           {isLoading && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
-          Reset Password
+          Sign In
         </Button>
       </form>
     </Form>
