+++ apps/web/app/(auth)/reset-password/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { Suspense } from "react";
-import { ResetPasswordForm } from "@workspace/ui/components/auth/reset-password-form";
+import { ResetPasswordForm } from "./reset-password-form";
 import {
   Card,
   CardContent,
