+++ apps/web/lib/constants.ts
@@ -1,5 +1,5 @@
 export const apiUrl =
-  process.env.NEXT_PUBLIC_API_URL || "http://localhost:8080/v1/";
+  process.env.NEXT_PUBLIC_API_URL || "http://localhost:8080/v1";
 
 export const appUrl =
   process.env.NEXT_PUBLIC_APP_URL || "http://localhost:3000";
