+++ apps/web/package.json
@@ -18,6 +18,9 @@
     "@radix-ui/react-slot": "^1.2.3",
     "@workspace/shared": "workspace:*",
     "@workspace/ui": "workspace:*",
+    "@hookform/resolvers": "^5.2.1",
+    "react-hook-form": "^7.62.0",
+    "sonner": "^2.0.7",
     "@zodios/core": "^10.9.6",
     "framer-motion": "^12.23.12",
     "lucide-react": "^0.475.0",
